% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_coolors_history.R
\name{clear_coolors_history}
\alias{clear_coolors_history}
\title{Clear coolors history from current session}
\usage{
clear_coolors_history(ask = TRUE)
}
\arguments{
\item{ask}{boolean; defaults to TRUE. Confirms deletion of coolors history if TRUE.}
}
\value{
Function intended for side-effects only.
}
\description{
Clear coolors history from current session
}
