% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolors_history.R
\name{coolors_history}
\alias{coolors_history}
\title{Display coolors palette history from current session}
\usage{
coolors_history()
}
\value{
A list of previously-generated coolors palettes. Limited to current session.
}
\description{
Display coolors palette history from current session
}
\examples{
coolors_history()
}
