% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_coolors.R
\name{palette_coolors}
\alias{palette_coolors}
\title{Coolors palette generator}
\usage{
palette_coolors(palette_num, n, hist = TRUE)
}
\arguments{
\item{palette_num}{integer; arbitrary palette identifier. Randomly generated by default, or can be specifically chosen from
existing Coolors palettes using an integer between 1 and 2092.}

\item{n}{integer; desired length of palette. Default length 5. Lengths greater than 5 will use color interpolation to
create palette from original 5-color palette.}

\item{hist}{logical; indicates whether an object storing palette history is created/updated for the current session. Since this
function's default color generation scheme is random, this  allows the user to review recently-generated palettes via their
identifiers, stored in a list. The most recent palette will be list item 1, the second-most recent list item 2, etc.}
}
\value{
A vector of colors.
}
\description{
Use \code{\link{palette_coolors}} to construct palettes of desired length
}
\examples{
# generate random palette
palette_coolors()

# random palette of length 10
palette_coolors(n = 10)

# if you liked a specific palette, specify its identifier
palette_coolors(palette_num = 1234, n = 20)

}
