% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_pal.R
\name{shuffle_pal}
\alias{shuffle_pal}
\title{Randomly shuffle part or all of a color palette}
\usage{
shuffle_pal(palette = .data, freeze, hist = TRUE)
}
\arguments{
\item{palette}{character vector of hexadecimal color values, with each beginning with "#". Alternatively, a
double vector of length 1 specifying the numeric identifier of \code{palette_coolors} palette.}

\item{freeze}{double vector specifying the positions of colors to freeze (i.e. not shuffle). Must be
length 1:n, where n is the length of the palette specified.}

\item{hist}{boolean, default TRUE; indicates whether to write results of shuffled palette to history
using \code{coolors_history()}.}
}
\value{
A vector of hexadecimal color values.
}
\description{
Randomly shuffle part or all of a color palette
}
