% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_hex.R
\name{valid_hex}
\alias{valid_hex}
\title{Test whether a string is a valid hexadecimal color value}
\usage{
valid_hex(hex)
}
\arguments{
\item{hex}{character vector containing string(s) to test as hex value(s).}
}
\value{
vector of boolean values indicating whether string(s) are valid hex code.
}
\description{
Test whether a string is a valid hexadecimal color value
}
\details{
A hex code is valid if:
\itemize{
\item It begins with '#' (can set \code{hash = FALSE} if needed), and
\item It is three, six, or eight characters long, consisting of the characters 0-9, A-F, and/or a-f.
}

Three-character hex codes are interpolated to six-character codes by repeating each element twice. Eight-
character hex codes contain two digits (either prepended or appended) representing transparency (alpha).
}
